requirements.txt
django==2.1.3
Markdown==3.0.1
djangorestframework==3.9.0
matplotlib==3.0.2
drf-writable-nested
# 连表更新插件
